Fix types

+++ xayn-ai-ffi-c/src/document.rs
@@ -247,7 +247,7 @@ impl ByteArray {
                 len: 0,
             }
         } else {
-            let len = bytes.len();
+            let len = bytes.len() as u32;
             let ptr = bytes.leak().as_mut_ptr();
 
             Self { ptr, len }
@@ -257,7 +257,7 @@ impl ByteArray {
     fn drop(array: *mut ByteArray) {
         if let Some(a) = unsafe { array.as_ref() } {
             if !a.ptr.is_null() && a.len > 0 {
-                unsafe { Box::from_raw(from_raw_parts_mut(a.ptr as *mut u8, a.len)) };
+                unsafe { Box::from_raw(from_raw_parts_mut(a.ptr as *mut u8, a.len as usize)) };
             }
             // Safety: we do not access `a` after we freed it
             unsafe { Box::from_raw(array) };
